Update GameScreen
Removed blank grid object (the object that was used to position the circle lower on the screen) and added a boxview object in its place. That way, we won't have a blank comment in the grid object which showed up as three dots on the actual gamescreen.

Also, there were two StackLayout tags but we only needed one so I removed that.