Bug fix
There was a bug where if the enemy position was at 355 but the player was on location 5 degrees, it looks like the player should be able to tap on the dot, but since 355 is far away from 5, the tap would not register. Also if the player was at 355 but the enemy was at 5, the same issue was present. I fixed this bug by ensuring the Enemy never spawns within 10 degrees of the "0" degree position